pyproject.toml
[project]
name = "doit-toolkit-cli"
version = "0.1.17"
description = "Doit CLI - A tool to bootstrap your projects for Spec-Driven Development (SDD)."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Sean Barlow", email = "sbarlow@barlowtech.com"}
]
requires-python = ">=3.11"
keywords = ["cli", "spec-driven-development", "sdd", "ai", "development-workflow", "scaffolding"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx[socks]",
    "platformdirs",
    "readchar",
    "truststore>=0.10.4",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/seanbarlow/doit"
Documentation = "https://github.com/seanbarlow/doit#readme"
Repository = "https://github.com/seanbarlow/doit.git"
Issues = "https://github.com/seanbarlow/doit/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-tmp-files>=0.0.2",
]
advanced = [
    "tiktoken>=0.5.0",
    "scikit-learn>=1.3.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.scripts]
doit = "doit_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doit_cli", "src/doit_toolkit_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "windows: tests that run only on Windows",
    "unix: tests that run only on Unix-like systems",
    "macos: tests that run only on macOS",
    "powershell: PowerShell script validation tests",
    "e2e: end-to-end tests",
    "filesystem: filesystem-specific tests",
    "bsd: tests BSD utility compatibility",
    "unicode: tests Unicode handling",
    "slow: tests that take longer than 10 seconds",
    "ci: tests that should run in CI",
]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = [
    ".venv",
    "src/doit_cli/templates",
    ".doit",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade — modernize type hints and syntax
    "B",    # flake8-bugbear — likely bugs and design issues
    "SIM",  # flake8-simplify — simplifiable constructs
    "PTH",  # flake8-use-pathlib — flag os.path leaks
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",    # line length handled by formatter
    "E402",    # module-import-not-at-top — Typer command modules use deferred imports
    "E741",    # ambiguous names (l, I, O) — small number of historical uses
    "B008",    # function call as default arg — Typer's Option/Argument use this pattern
    "PTH",     # os.path → pathlib — addressed in Phase 3 (service refactor)
    "SIM102",  # collapsible-if — case-by-case readability call
    "SIM105",  # suppressible-exception — wholesale contextlib.suppress can hide intent
    "SIM108",  # ternary rewrites hurt readability for branching logic
    "RUF001",  # ambiguous unicode — a handful of em-dashes in docstrings, intentional
    "RUF012",  # mutable class defaults — addressed in Phase 3 (model hardening)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B011", "B017", "SIM117"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
# Start permissive; tighten in a follow-up phase.
disallow_untyped_defs = false
check_untyped_defs = true
exclude = [
    "src/doit_cli/templates",
    "tests",
    ".venv",
]

# Modules excluded from mypy enforcement. Keep this list minimal — every
# entry is tech debt. Prefer fixing types in-file over extending this.
[[tool.mypy.overrides]]
module = [
    # Legacy services not yet cleaned up by the modernization effort.
    # Remove each entry after its module has been properly typed.
    "doit_cli.services.template_manager",
    "doit_cli.services.github_service",
    "doit_cli.services.wizard_service",
    "doit_cli.services.milestone_service",
    "doit_cli.services.memory_search",
    "doit_cli.cli.roadmapit_impl",
    "doit_cli.mcp.tools.status_tool",
]
ignore_errors = true

